sw/ground_segment/python/no_fly_zones/requirements.txt
asterix_decoder==0.5.5